requirements.txt
paho-mqtt==1.6.1
pymodbus==3.5.4
pyserial==3.5 